add deeplabel tool to readme
I'm the author of deeplabel, an open source bounding box annotation tool. I saw some other labelling tools are in the readme and hope you'd consider adding this one as well - it's a GUI app written in C++, binaries are provided for Windows and Mac, and it compiles fine on Linux. Deeplabel imports from and exports directly to darknet format and also allows users to load darknet weights for automatically labelling an image dataset (via the OpenCV backend) - though this currently only works up to yolov3. I'm hoping in the near future to also add support for adding train support via a user-supplied darknet binary, or at least generating the config files automatically.